Added non-nested layout system for all social shortcodes
