+++ layouts/social-feed/waterfall.php
@@ -1,0 +1,6 @@
+<?php
+
+/**
+ * Alias for the default layout.
+ **/
+add_filter( 'ucf_social_feed_display_waterfall', 'ucf_social_feed_display_default', 10, 2 );
