+++ ucf-social.php
@@ -15,11 +15,20 @@ add_action( 'plugins_loaded', function() {
 
 	define( 'UCF_SOCIAL__PLUGIN_FILE', __FILE__ );
 
-	require_once 'layouts/layout-social-feed-grid.php';
-	require_once 'layouts/layout-social-feed-waterfall.php';
+	// Layouts - social feed
+	require_once 'layouts/social-feed/grid.php';
+	require_once 'layouts/social-feed/waterfall.php';
+
+	// Layouts - social icons
+	require_once 'layouts/social-icons/default.php';
+
+	// Layouts - social links
+	require_once 'layouts/social-links/default.php';
+
+	// Core plugin files
 	require_once 'includes/ucf-social-config.php';
 	require_once 'includes/ucf-social-common.php';
-	require_once 'includes/ucf-social-shortcode.php';
+	require_once 'includes/ucf-social-shortcodes.php';
 
 } );
